Refactored and commented code, for clarity.
Should not be any functional changes here.

BFD-1071
